distributed_infra_env/server/requirements.txt
openenv-core[core]>=0.2.2
fastapi>=0.115.0
pydantic>=2.0.0
uvicorn>=0.24.0
requests>=2.31.0
